chore: move existing test files to __tests__ dir
