+++ packages/contract-helpers/src/gho/__tests__/index.test.ts
@@ -1,0 +1,19 @@
+import {
+  GhoDiscountRateStrategyService,
+  GhoTokenService,
+  GhoVariableDebtTokenService,
+} from '../index';
+
+describe('gho contract helpers', () => {
+  it('exports out the GhoDiscountRateStrategyService', () => {
+    expect(GhoDiscountRateStrategyService).toBeTruthy();
+  });
+
+  it('exports out the GhoTokenService', () => {
+    expect(GhoTokenService).toBeTruthy();
+  });
+
+  it('exports out the GhoVariableDebtTokenService', () => {
+    expect(GhoVariableDebtTokenService).toBeTruthy();
+  });
+});
