--- packages/contract-helpers/src/gho/index.test.ts
@@ -1,7 +1,0 @@
-import { GhoDiscountRateStrategyService } from './index';
-
-describe('gho contract helpers', () => {
-  it('exports out the GhoDiscountRateStrategyService', () => {
-    expect(GhoDiscountRateStrategyService).toBeTruthy();
-  });
-});
