+++ packages/contract-helpers/src/gho/index.ts
@@ -1,3 +1,9 @@
 import { GhoDiscountRateStrategyService } from './GhoDiscountRateStrategyService';
+import { GhoTokenService } from './GhoTokenService';
+import { GhoVariableDebtTokenService } from './GhoVariableDebtTokenService';
 
-export { GhoDiscountRateStrategyService };
+export {
+  GhoDiscountRateStrategyService,
+  GhoTokenService,
+  GhoVariableDebtTokenService,
+};
